Retab and rename
