+++ src/test/java/models/testEnglishGame.java
@@ -1,0 +1,22 @@
+package models;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+public class testEnglishGame{
+
+    @Test
+    public void testEnglishGameCreation(){
+        Game s = new englishGame();
+        assertNotNull(s);
+    }
+
+    @Test
+    public void testEnglishDeckBuilt(){
+        Game s = new englishGame();
+        s.buildDeck();
+        assertEquals(52, s.deck.size());
+    }
+
+}
