--- src/test/java/models/testenglishGame.java
@@ -1,22 +1,0 @@
-package models;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-public class testEnglishGame{
-
-  @Test
-  public void testEnglishGameCreation(){
-    Game s = new englishGame();
-    assertNotNull(s);
-  }
-
-  @Test
-  public void testEnglishDeckBuilt(){
-    Game s = new englishGame();
-    s.buildDeck();
-    assertEquals(52, s.deck.size());
-  }
-
-}
